Add erlang:nif_error/1,2
A stub function that is supposed to be replaced by a NIF usually
calls erlang:error/1 to cause an exception if the NIF library
is not loaded. For example:

  foo() ->
      erlang:error(nif_not_loaded).

The problem is that although erlang:error/1 will normally never be
called, Dialyzer will think that any call to the function will fail
and thus generate false warnings. Adding a spec for the function
will not help because Dialyzer will not believe the spec.

Add erlang:nif_error/1,2 that work exactly like erlang:error/1,2.

Define the return types for both BIFs to be t_any().

erlang:nif_error is used like this:

  -spec foo() -> binary().

  foo() ->
      erlang:nif_error(nif_not_loaded).

(The -spec is optional but highly recommended, since Dialyzer
otherwise has no chance to figure out the types.)