Merge branch 'ks/dialyzer-R14-fixes' into dev
* ks/dialyzer-R14-fixes:
  Various changes to dialyzer-related files for R14.

OTP-8699 ks/dialyzer-R14-fixes

Various changes to dialyzer-related files for R14.

- Dialyzer properly supports the new attribute -export_type and checks that
remote types only refer to exported types. A warning is produced if some
files/applications refer to types defined in modules which are neither in
the PLT nor in the analyzed applications.

- Support for detecting data races involving whereis/1 and unregister/1.

- More precise identification of the reason(s) why a record construction
violates the types declared for its fields.

- Fixed bug in the handling of the 'or' guard.

- Better handling of the erlang:element/2 BIF.

- Complete handling of Erlang BIFs.