hipe: Remove single function compilation
Compiling single functions (as opposed to whole modules) to native
code complicates code management in HiPE. It would also vastly
complicate whole-module optimizations, such as returning multiple
return values instead of tuples within a module.

As a first step, remove the external interface for the single
compilation feature. In the future, there are many things that
could be restructured and simplified.